# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aftersales Api - SP21",
        "description": "List Aftersales Api",
        "version": "1.0"
    },
    "paths": {
...
        "/product/{id}/comment": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Show comments of the product has id = $id",
                "operationId": "b08c071dda1c08be7121671ce10122a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "product_id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": 2
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "example comment"
                                                    },
                                                    "photo": {
                                                        "type": "string",
                                                        "example": "null"
                                                    },
                                                    "created_at": {
                                                        "type": "timestamp",
                                                        "example": "2022-01-20T09:06:33.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "timestamp",
                                                        "example": "2022-01-20T09:06:33.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create a new comment",
                "operationId": "69597642280b87ae1c0b4abb2b2db909",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "user_id",
                                    "comment",
                                    "photo"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "user_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "comment": {
                                        "type": "text",
                                        "example": "Sản phẩm đẹp quá"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "example": "null"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Success": {
                                            "type": "string",
                                            "example": "Send a comment successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comment/{id}": {
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Delete a comment by comment's id",
                "description": "Delete a comment",
                "operationId": "9bdf98bdab49153f26c620e7e975ecab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a comment successfully"
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}